requirements.txt
pygame-ce
opencv-python
mediapipe
SpeechRecognition
pyaudio